+++ store/v2/commitment/store.go
@@ -12,7 +12,8 @@ import (
 	corestore "cosmossdk.io/core/store"
 	"cosmossdk.io/log"
 	"cosmossdk.io/store/v2"
-	internal "cosmossdk.io/store/v2/internal/conv"
+	"cosmossdk.io/store/v2/internal"
+	"cosmossdk.io/store/v2/internal/conv"
 	"cosmossdk.io/store/v2/internal/encoding"
 	"cosmossdk.io/store/v2/proof"
 	"cosmossdk.io/store/v2/snapshots"
@@ -40,7 +41,7 @@ type CommitStore struct {
 	multiTrees map[string]Tree
 
 	// pruneOptions is the pruning configuration.
-	pruneOptions *store.PruneOptions
+	pruneOptions *store.PruneOptions // TODO are there no default prune options?
 }
 
 // NewCommitStore creates a new CommitStore instance.
@@ -60,7 +61,7 @@ func NewCommitStore(trees map[string]Tree, db corestore.KVStoreWithBatch, pruneO
 func (c *CommitStore) WriteBatch(cs *corestore.Changeset) error {
 	for _, pairs := range cs.Changes {
 
-		key := internal.UnsafeBytesToStr(pairs.Actor)
+		key := conv.UnsafeBytesToStr(pairs.Actor)
 
 		tree, ok := c.multiTrees[key]
 		if !ok {
@@ -83,6 +84,9 @@ func (c *CommitStore) WriteBatch(cs *corestore.Changeset) error {
 func (c *CommitStore) WorkingCommitInfo(version uint64) *proof.CommitInfo {
 	storeInfos := make([]proof.StoreInfo, 0, len(c.multiTrees))
 	for storeKey, tree := range c.multiTrees {
+		if internal.IsMemoryStoreKey(storeKey) {
+			continue
+		}
 		bz := []byte(storeKey)
 		storeInfos = append(storeInfos, proof.StoreInfo{
 			Name: bz,
@@ -198,6 +202,9 @@ func (c *CommitStore) Commit(version uint64) (*proof.CommitInfo, error) {
 	storeInfos := make([]proof.StoreInfo, 0, len(c.multiTrees))
 
 	for storeKey, tree := range c.multiTrees {
+		if internal.IsMemoryStoreKey(storeKey) {
+			continue
+		}
 		// If a commit event execution is interrupted, a new iavl store's version
 		// will be larger than the RMS's metadata, when the block is replayed, we
 		// should avoid committing that iavl store again.
@@ -251,7 +258,7 @@ func (c *CommitStore) SetInitialVersion(version uint64) error {
 }
 
 func (c *CommitStore) GetProof(storeKey []byte, version uint64, key []byte) ([]proof.CommitmentOp, error) {
-	tree, ok := c.multiTrees[internal.UnsafeBytesToStr(storeKey)]
+	tree, ok := c.multiTrees[conv.UnsafeBytesToStr(storeKey)]
 	if !ok {
 		return nil, fmt.Errorf("store %s not found", storeKey)
 	}
@@ -277,7 +284,7 @@ func (c *CommitStore) GetProof(storeKey []byte, version uint64, key []byte) ([]p
 }
 
 func (c *CommitStore) Get(storeKey []byte, version uint64, key []byte) ([]byte, error) {
-	tree, ok := c.multiTrees[internal.UnsafeBytesToStr(storeKey)]
+	tree, ok := c.multiTrees[conv.UnsafeBytesToStr(storeKey)]
 	if !ok {
 		return nil, fmt.Errorf("store %s not found", storeKey)
 	}
@@ -376,7 +383,12 @@ func (c *CommitStore) Snapshot(version uint64, protoWriter protoio.Writer) error
 }
 
 // Restore implements snapshotstypes.CommitSnapshotter.
-func (c *CommitStore) Restore(version uint64, format uint32, protoReader protoio.Reader, chStorage chan<- *corestore.StateChanges) (snapshotstypes.SnapshotItem, error) {
+func (c *CommitStore) Restore(
+	version uint64,
+	format uint32,
+	protoReader protoio.Reader,
+	chStorage chan<- *corestore.StateChanges,
+) (snapshotstypes.SnapshotItem, error) {
 	var (
 		importer     Importer
 		snapshotItem snapshotstypes.SnapshotItem